Fix deployment fastfetch

+++ install/desktop/fonts.sh
@@ -1,22 +1,43 @@
 #!/bin/bash
 
-mkdir -p ~/.local/share/fonts
+set -e
+
+FONT_DIR="$HOME/.local/share/fonts"
+
+mkdir -p "$FONT_DIR"
 
 cd /tmp
-wget https://github.com/ryanoasis/nerd-fonts/releases/latest/download/CascadiaMono.zip
-unzip CascadiaMono.zip -d CascadiaFont
-cp CascadiaFont/*.ttf ~/.local/share/fonts
-rm -rf CascadiaMono.zip CascadiaFont
 
-wget -O iafonts.zip https://github.com/iaolo/iA-Fonts/archive/refs/heads/master.zip
-unzip iafonts.zip -d iaFonts
-cp iaFonts/iA-Fonts-master/iA\ Writer\ Mono/Static/iAWriterMonoS-*.ttf ~/.local/share/fonts
-rm -rf iafonts.zip iaFonts
+# CASCADIA MONO
+if ! find "$FONT_DIR" -iname "CaskaydiaMono*.ttf" -o -iname "CascadiaMono*.ttf" | grep -q .; then
+  wget -q -O CascadiaMono.zip \
+    https://github.com/ryanoasis/nerd-fonts/releases/latest/download/CascadiaMono.zip
+  unzip -q -o CascadiaMono.zip -d CascadiaFont
+  cp CascadiaFont/*.ttf "$FONT_DIR"/
+  rm -rf CascadiaMono.zip CascadiaFont
+fi
 
-wget -O JetBrainsMono.zip https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.zip
-unzip -o JetBrainsMono.zip -d JetBrainsMono
-cp JetBrainsMono/*.ttf ~/.local/share/fonts
-rm -rf JetBrainsMono.zip JetBrainsMono
+# IA WRITER MONO
+if ! find "$FONT_DIR" -iname "iAWriterMonoS-*.ttf" | grep -q .; then
+  wget -q -O iafonts.zip \
+    https://github.com/iaolo/iA-Fonts/archive/refs/heads/master.zip
+  unzip -q -o iafonts.zip -d iaFonts
+  cp iaFonts/iA-Fonts-master/iA\ Writer\ Mono/Static/iAWriterMonoS-*.ttf \
+    "$FONT_DIR"/
+  rm -rf iafonts.zip iaFonts
+fi
 
-fc-cache
+# JETBRAINS MONO
+if ! find "$FONT_DIR" -iname "JetBrainsMono*.ttf" | grep -q .; then
+  echo "Installing JetBrains Mono Nerd Font..."
+  wget -q -O JetBrainsMono.zip \
+    https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.zip
+  unzip -q -o JetBrainsMono.zip -d JetBrainsMono
+  cp JetBrainsMono/*.ttf "$FONT_DIR"/
+  rm -rf JetBrainsMono.zip JetBrainsMono
+fi
+
+# REFRESH FONT CACHE
+fc-cache -f
+
 cd -
